pickgok-ai-server/requirements.txt
# --- Core Framework ---
flask>=2.3.0

# --- AI & Vector Search ---
faiss-cpu==1.13.0
numpy>=1.24.0
scikit-learn>=1.3.0

# --- Audio Processing (Only for build_index.py) ---
librosa==0.10.1

# --- Data Handling ---
pandas>=2.0.0
tqdm==4.66.3

# --- Database (New) ---
# Python에서 MySQL로 메타데이터를 밀어넣기 위해 필요
mysql-connector-python>=8.0.0